Multiplayer games did not work on 64 bit Linux
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@356 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/network.h
@@ -170,7 +170,7 @@ typedef struct tExtraGameInfo {
 	char		bFriendlyFire;
 	char		bInhibitSuicide;
 	char		bFixedRespawns;
-	long		nSpawnDelay;
+	int		nSpawnDelay;
 	char		bEnhancedCTF;
 	char		bRadarEnabled;
 	char		bPowerupsOnRadar;
